[Structs] Add Bot struct
Add a bot struct. This is just a generic struct for thinking of the bots as a separate entity.

We'll actually try to do some polymorphism here and make the
bot a "GridEntity" (and hopefully mining nodes also will be this later).

Add a basic struct for bots that takes an ID and a list of claims.
Make it part of the registration process.