[Server] Semi-major refactor
Couldn't get the mux tests to work properly.

There were a lot of problems with how I was handling things,
and in general this architecture seems much cleaner to me.

POST body handling seems somewhat fragile in Go,
but nonetheless this seems to work
and tests are passing again properly.