[Testing] Add more thorough tests, add error conditions
1. Add more thorough tests

We had left some tests incomplete, so in the process
of adding new error conditions let's finish these up.

2. Add error conditions

We noted these previously, but basically you shouldn't
be allowed to claim or mine a node if you're too far away.

Also we have a maximum number of allowable claims, so enforce that.